# api/docs/swagger.yaml
definitions:
...
  models.CategorySwagger:
    properties:
      id:
        example: a2d70daf-b4ac-4198-a6a0-999447483c18
        type: string
      image:
        example: https://firebasestorage.googleapis.com/v0/b/ulab-market.appspot.com/o/test%2Fcategory%2Fa2d70daf-b4ac-4198-a6a0-999447483c18?alt=media&token=test%2Fcategory%2Fa2d70daf-b4ac-4198-a6a0-999447483c18
        type: string
      name:
        example: electronic
        type: string
      parent_id:
        example: ""
        type: string
      subcategories:
        items:
          $ref: '#/definitions/models.SubCategorySwagger'
        type: array
      translations:
        items:
          $ref: '#/definitions/models.CategoryTranslation'
        type: array
    type: object
  models.CategoryTranslation:
    properties:
      category_id:
        type: string
      language_code:
        type: string
      name:
        type: string
    type: object
...
  models.SubCategorySwagger:
    properties:
      id:
        type: string
      image:
        type: string
      name:
        type: string
      parent_id:
        type: string
      translations:
        items:
          $ref: '#/definitions/models.CategoryTranslation'
        type: array
    type: object
  models_v1.CategoryTranslation:
    properties:
      category_id:
        type: string
      language_code:
        type: string
      name:
        type: string
    required:
    - category_id
    - language_code
    - name
    type: object
...
  models_v1.ChangeCategory:
    properties:
      id:
        type: string
      name:
        type: string
      parent_id:
        type: string
    required:
    - id
    - name
    type: object
...
  models_v1.CreateCategory:
    properties:
      name:
        type: string
      parent_id:
        type: string
    required:
    - name
    type: object
...
  models_v1.Response:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  models_v1.ResponseID:
    properties:
      id:
        type: string
    type: object
...
host: localhost:8123
info:
  contact: {}
  title: E-commerce project v2
paths:
...
  /api/category:
    get:
      consumes:
      - application/json
      description: get category, returns translations, and subcategories for all category
      operationId: getAllCategory
      produces:
      - application/json
      responses:
        "200":
          description: category returned
          schema:
            items:
              $ref: '#/definitions/models.CategorySwagger'
            type: array
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/models_v1.Response'
      summary: get category all
      tags:
      - category
    post:
      consumes:
      - application/json
      description: Create category
      operationId: createCategory
      parameters:
      - description: Create category request
        in: body
        name: create_category
        required: true
        schema:
          $ref: '#/definitions/models_v1.CreateCategory'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/models_v1.ResponseID'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models_v1.Response'
        "409":
          description: Already exists
          schema:
            $ref: '#/definitions/models_v1.Response'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/models_v1.Response'
      security:
      - ApiKeyAuth: []
      summary: Create category
      tags:
      - category
    put:
      consumes:
      - application/json
      description: change category name and parent
      operationId: changeCategory
      parameters:
      - description: change category
        in: body
        name: changeCategory
        required: true
        schema:
          $ref: '#/definitions/models_v1.ChangeCategory'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/models_v1.Response'
        "400":
          description: Bad request / bad uuid
          schema:
            $ref: '#/definitions/models_v1.Response'
        "404":
          description: Category not found/ parent category not found
          schema:
            $ref: '#/definitions/models_v1.Response'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/models_v1.Response'
      security:
      - ApiKeyAuth: []
      summary: change category
      tags:
      - category
  /api/category/{id}:
    delete:
      consumes:
      - application/json
      description: delete category & delete category translations
      operationId: deleteCategory
      parameters:
      - description: category id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: deleted successfully
          schema:
            $ref: '#/definitions/models_v1.Response'
        "400":
          description: bad uuid
          schema:
            $ref: '#/definitions/models_v1.Response'
        "404":
          description: category not found
          schema:
            $ref: '#/definitions/models_v1.Response'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/models_v1.Response'
      security:
      - ApiKeyAuth: []
      summary: delete category
      tags:
      - category
    get:
      consumes:
      - application/json
      description: get category by id, returns translations, and subcategories for
        specified category
      operationId: getCategoryById
      parameters:
      - description: category id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: category returned
          schema:
            $ref: '#/definitions/models.CategorySwagger'
        "400":
          description: Bad UUID
          schema:
            $ref: '#/definitions/models_v1.Response'
        "404":
          description: Category not found
          schema:
            $ref: '#/definitions/models_v1.Response'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/models_v1.Response'
      summary: get category by id
      tags:
      - category
  /api/category/add_translation:
    post:
      consumes:
      - application/json
      description: Create category translation
      operationId: AddCategoryTranslation
      parameters:
      - description: Create category translation request
        in: body
        name: create_category_translation
        required: true
        schema:
          $ref: '#/definitions/models_v1.CategoryTranslation'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/models_v1.Response'
        "400":
          description: Bad request/ Bad id
          schema:
            $ref: '#/definitions/models_v1.Response'
        "404":
          description: Category not found
          schema:
            $ref: '#/definitions/models_v1.Response'
        "409":
          description: Already exists
          schema:
            $ref: '#/definitions/models_v1.Response'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/models_v1.Response'
      security:
      - ApiKeyAuth: []
      summary: Create category translation
      tags:
      - category
  /api/category/change_image:
    post:
      consumes:
      - application/json
      description: change category image
      operationId: ChangeCategoryImage
      parameters:
      - in: formData
        name: category_id
        required: true
        type: string
      - description: picture file
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/models_v1.Response'
        "400":
          description: Bad request / bad uuid
          schema:
            $ref: '#/definitions/models_v1.Response'
        "404":
          description: Category not found
          schema:
            $ref: '#/definitions/models_v1.Response'
        "413":
          description: Image size is big
          schema:
            $ref: '#/definitions/models_v1.Response'
        "415":
          description: Image type is not supported
          schema:
            $ref: '#/definitions/models_v1.Response'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/models_v1.Response'
      security:
      - ApiKeyAuth: []
      summary: change category image
      tags:
      - category
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"